+++ CHANGES.rst
@@ -89,6 +89,7 @@ Improvements
   view (:issue:`4513`)
 - Add "Add myself" button to person list fields (e.g. for abstract authors)
   (:issue:`4411`, thanks :user:`jgrigera`)
+- Add CfA setting to control whether authors can edit abstracts (:issue:`3431`)
 
 Bugfixes
 ^^^^^^^^